pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "flintcore-testing-platform"
version = "0.1.0"
description = "Regression and CI harness for the FlintCore chess engine"
readme = "README.md"
authors = [{name = "FlintCore Maintainers"}]
requires-python = ">=3.10"
dependencies = [
    "python-chess>=1.999",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
]

[tool.setuptools]
packages = ["flintcoretest"]
include-package-data = true

[tool.setuptools.package-data]
flintcoretest = ["data/*.epd"]
